backend/requirements.txt
fastapi
uvicorn
pydantic
openai
motor
python-multipart
python-dotenv
pytest
